requirements_new.txt
numpy
opencv-python-headless==4.10.0.84
requests
requests-toolbelt==1.0.0
roboflow==1.2.11
pyserial==3.4
pydobot @ git+https://github.com/luismesas/pydobot.git@2ff8d7e1318ac9b7bc32bd33ff327d343f44927c
 